Adding my SQL script file from class
In class, the instructor told us to add everything to one sql file. The create database command, use database command, create tables, insert data, etc. are all in this one file. This seems like not a great way to do it because things have to be commented out in order to run the file. Moving forward, I may separate this into multiple files.